Revise the declaring/primary constructor proposal (#4480)
Revise the declaring/primary constructor proposal according to the language meeting Jul 31.

In particular, change the declaring/primary constructor proposal to use a different scoping structure (where initializing expressions of non-late instance variables have access to the formal parameters of primary constructor parameters, if there is a primary constructor); clarify the placement of the modifier const with a declaring constructor, to mention the principle that the class header should contain all parts of the signature of a declaring constructor if it contains any of them; perform several other minor clarifications and corrections.