Update the declaring constructors feature specification (#4524)
This PR makes several small corrections to the declaring constructor feature specification. It also renames the document (and the feature) to 'declaring constructors', which would remove some confusion about the word 'primary': A primary constructor is one of the kinds of declaring constructors that we can create.

Finally, this PR also introduces support for an abbreviated syntax with non-declaring constructors (not because it's a necessary part of the declaring constructors feature, but because it fits well with the emphasis on brevity):

```dart
// Previously:
class LooooooooooooooooooooooooooooooooongName {
  const LooooooooooooooooooooooooooooooooongName();
  LooooooooooooooooooooooooooooooooongName.surName(): this();
}

// Can now be:
class LooooooooooooooooooooooooooooooooongName {
  const new();
  new.surName(): this();
}
```