Change `classMemberDeclaration` to `memberDeclaration`
The grammar in the language specification used to have a non-terminal `<classMemberDeclaration>`, and the spec parser had special variants like `<mixinMemberDeclaration>` and `<extensionTypeMemberDeclaration>`. This would be useful if different membered declarations have different syntax, and that was once the plan.

However, we never actually got around to make this kind of distinction, and it's a more useful approach to report disallowed kinds of member declarations using ad-hoc compile-time checks (which is also what the implementations do already).

This PR changes all references to these non-terminals to the plain name `<memberDeclaration>` and adjusts the accompanying explanations accordingly.