Add motivation to "Augmentations" about code generation and FFI.
We usually think of augmentations as being generated code augmenting
hand-authored code. But there are use cases in the other direction as
well where the code generator produces the base declaration and a user
hand-writes an augmentation to tweak it.

Add some verbiage to the motivation section to capture that.