Add rules to allow omission of certain default values (#4393)
This PR changes the rules about augmentation of constructors such that an incomplete factory is able to defer the decision about implementing the constructor by a redirection or by a body. This is done by making the rules about default values of formal parameters more flexible: The default value can be specified in multiple locations as long as there is at most one of them.