Update declaring constructors to allow the new constructor syntax (#4543)
Update the declaring constructors feature specification to allow the new constructor syntax (`new();` rather than `ClassName();`, etc). This gives rise to an ambiguity in the grammar (`factory() => e;` can be a method named `factory` and no return type, but it can also be a factory constructor whose name is the class name; it's resolved to make it a factory constructor).