Add sections about reading the code and flow analysis (#4554)
This commit adds two sections to the feature specification about anonymous methods. First, the introduction is extended to give some ideas about how the code can be read aloud, yielding some hints about the semantics. Next, the proposal is extended with a section that briefly describes the flow analysis of an anonymous method invocation expression.