Specify that non-plain primary parameters are immutable (#4575)
Specify that primary parameters are immutable in initializer expressions (but a primary parameter can still be modified in the function body of a primary constructor). Also specify that double initialization is an error with a primary constructor (initializing formal + initializing expression in instance variable declaration + initializer list, or two of those).